update bot for bounty3 which is now just bounty
