+++ bin/runtime/Cargo.toml
@@ -79,5 +79,4 @@ treasury = { package = "sunshine-treasury", path = "../../pallets/treasury", def
 donate = { package = "sunshine-donate", path = "../../pallets/donate", default-features=false}
 bank = { package = "sunshine-bank", path = "../../pallets/bank", default-features=false}
 bounty = { package = "sunshine-bounty", path = "../../pallets/bounty", default-features=false}
-bounty3 = { package = "sunshine-bounty3", path = "../../pallets/bounty3", default-features=false}
 serde = { version = "1.0.114", features = ["derive"], optional = true }
