+++ bin/runtime/src/lib.rs
@@ -415,19 +415,11 @@ impl bank::Trait for Runtime {
     type MinimumInitialDeposit = MinimumInitialDeposit;
 }
 parameter_types! {
-    pub const BountyLowerBound: u128 = 5;
-}
-impl bounty::Trait for Runtime {
-    type Event = Event;
-    type BountyId = u64;
-    type BountyLowerBound = BountyLowerBound;
-}
-parameter_types! {
     pub const Foundation: ModuleId = ModuleId(*b"fundacon");
     pub const MinDeposit: u128 = 10;
     pub const MinContribution: u128 = 5;
 }
-impl bounty3::Trait for Runtime {
+impl bounty::Trait for Runtime {
     type Event = Event;
     type IpfsReference = CidBytes;
     type Currency = Balances;
@@ -462,7 +454,6 @@ construct_runtime!(
         Donate: donate::{Module, Call, Event<T>},
         Bank: bank::{Module, Call, Storage, Event<T>},
         Bounty: bounty::{Module, Call, Storage, Event<T>},
-        Bounty3: bounty3::{Module, Call, Storage, Event<T>},
     }
 );
