+++ client/cli/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod bank;
 pub mod bounty;
-pub mod bounty3;
 pub mod donate;
 mod error;
 pub mod org;
