--- pallets/bounty3/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "sunshine-bounty3"
-version = "0.0.1"
-authors = ["Amar Singh <asinghchrony@protonmail.com>"]
-edition = "2018"
-
-license = "GPL-3.0"
-repository = "https://github.com/sunshine-protocol/sunshine-bounty"
-description = "minimal sunshine-bounty module"
-keywords = ["sunshine", "substrate"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-sp-std = { version = "2.0.0-rc5", default-features = false }
-sp-runtime = { version = "2.0.0-rc5", default-features = false }
-frame-support = { version = "2.0.0-rc5", default-features = false }
-frame-system = { version = "2.0.0-rc5", default-features = false }
-util = { package = "sunshine-bounty-utils", path = "../../utils", default-features=false}
-clear_on_drop = { version = "0.2.4", features = ["no_cc"] }	# https://github.com/paritytech/substrate/issues/4179
-
-[dev-dependencies]
-serde = { version = "1.0.114" }
-sp-io = { version = "2.0.0-rc5", default-features = false }
-sp-core = { version = "2.0.0-rc5", default-features = false }
-pallet-balances = { version = "2.0.0-rc5", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-	"codec/std",
-	"sp-std/std",
-	"sp-runtime/std",
-	"frame-support/std",
-	"frame-system/std",
-	"pallet-balances/std",
-]
